task2

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -9,7 +9,7 @@
 int get_bit(unsigned long int n, unsigned int index)
 {
 	unsigned int i = 0;
-	int l;
+	unsigned long int l;
 
 	if ((n == 0 || n == 1) && index == 0)
 		return (n);
